iwlwifi: mvm: add the quota remainder to a data binding
Currently the quota remainder was added to the first binding, although
it is possible that this was not a data binding (only the P2P_DEVICE
interface is part of the binding).

Fix this by adding the remainder to the first binding that was actually
allocated quota.

Change-Id: Ibb62598d64196a19a5591cd881dbb36e62f78787
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/29174
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: 9f5e8a9950dd302fc4c42ad44f4b8e23aee7b167