cfg80211: send scan results from work queue [BUGFIX]
Due to the previous commit, when a scan finishes, it is in theory
possible to hit the following sequence:
 1. interface starts being removed
 2. scan is cancelled by driver and cfg80211 is notified
 3. scan done work is scheduled
 4. interface is removed completely, rdev->scan_req is freed,
    event sent to userspace but scan done work remains pending
 5. new scan is requested on another virtual interface
 6. scan done work runs, freeing the still-running scan

To fix this situation, hang on to the scan done message and block
new scans while that is the case, and only send the message from
the work function, regardless of whether the scan_req is already
freed from interface removal. This makes step 5 above impossible
and changes step 6 to be
 5. scan done work runs, sending the scan done message

As this can't work for wext, so we send the message immediately,
but this shouldn't be an issue since we still return -EBUSY.

Change-Id: I35d9d4cf1b8569e0541ce204fbf2be04afb3f074
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
---
It seems practically impossible to hit this bug without changing
the code to make it more likely. It may be easier to hit on single-
CPU systems (hello xmm6321) but I'm not sure.
Reviewed-on: https://gerrit.rds.intel.com/29259
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: 6995732fc50bf4220e07fd86eb673d535ce2787d