iwlwifi: mvm: BT Coex - fix SYNC2SCO flags
The Sync to SCO is a feature that allows to synchronize
between the WiFi traffic and the expectable BT traffic
when SCO profile is active.
We need to set the validity bit in the command in the init
flow, and set / clear the enablement bit if we want to
enabled / disable the feature.

While at it, clean up the flags that are not used in the
API.

Change-Id: Ic5e43ca27913c513dac6883acf3b2bdefb167a69
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/29789
Tested-by: IWL Jenkins
Reviewed-by: Eyal Zolotov <eyal.zolotov@intel.com>
Tested-by: Eyal Zolotov <eyal.zolotov@intel.com>

x-iwlwifi-stack-dev: 77aa5df0016b397136d2b13a078702bea8a8d973