iwlwifi: remove potential buffer overrun in fw name
This patch fixes a potential buffer overrun when creating
the fw name in drv->firmware_name by setting a maximal
length to the char array copied to it. The maximal length
is also updated to 32 rather than 25 to keep both 32bit and
64bit alignment without requiring padding to the struct it
is in.

Change-Id: If4cc38fef65f27c1e20d2db9b872c4e1313b4afc
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/29832
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: c9640ef78ec32e293a7b084769b3c9d6fa7aa973