iwlwifi: mvm: refactor power code [BUGFIX]
The main complexity of the power code is that it needs to
take into account the firmware limitations.
These limitations state that we need to have global picture
of the vifs present in the system to be able to decide if
we can enable power management on a specific vif.
Even device power save (as opposed to vif power management)
must be disabled in certain circumstances (monitor vif).

Refactor the current code to make this clearer by defining
a function that explicitely computes these constraints.

Change-Id: Ie365d22d15c3979a97c891d5273ebc43c5c8d835
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
---
MWG100193871
Reviewed-on: https://gerrit.rds.intel.com/29831
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: 098a4ae678ca0b5c3227e83458e6edff0919075e