iwlwifi: mvm: modify the tsf_id master/slave logic [BUGFIX]
For TSF master/slave synchronization, the FW does not require
exact match in the beacon interval between the master interface
and the slave one, but instead requires that the beacon interval
of one interface is the module of the other.

Modify the tsf_id selection to align with the above.

---

This part of the fix of MWG100197066 (also requires hostap changes)

Change-Id: Ic349df85ab8bd9f3257aaee3ffd39461e5bea6ce
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/30265
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 45ec08f78876c38ade2b84f283b7e52e2e028f29