iwlwifi: always call opmode RX with BHs disabled
This should probably be squashed into the NAPI patch for
upstream, since that patch introduced the issue.

The issue is that we sometimes call into the opmode with
BHs enabled and sometimes disabled, which means it can't
rely on either way, which is less efficient.

Additionally, it seems to me now that the NAPI prepare
should be done with BHs disabled, since otherwise it would
seem possible that it changed, though I'm not sure.

Change-Id: I5542ff0e7f9b8bde43d979fcb590139d9e1374f8
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/30614
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: ab878928b1538d165e07830aa6e07d4b1dc9758a