iwlwifi: mvm: abort scan on sched_scan request
A scheduled scan is a more persistent setting and should take priority
over temporary regular scans. Abort the regular when a sched_scan
request arrives and then request the sched_scan.

The kernel API allows sending a sched_scan without canceling a regular
scan in progress, so this is our way to abstract the FW's limitations.

Make the scan-cancel Rx handler async and flush after invocation to
ensure new scans can't creep in before it.

--
This is needed for Android PNO which uses sched scan. We don't recover
from a failed attempt, which is bad since we're suspending the host. But
it's a good practice in general as well.

Change-Id: I99da7e6a28ad25819f99f07189090ef886f91e17
Signed-off-by: Arik Nemtsov <arikx.nemtsov@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/30743
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: 35243393b299d8d22a6aa16e4650f6c9ef7f7965