iwlwifi: pcie: stop the firmware when we restart it
In case the firmware didn't assert but we want to restart
it, e.g. we didn't get the reply for a host command, or the
Tx queues are stuck, we should stop the firmware by
provoking an interrupt. This allows to better debug the
firmware in these bad scenarios.

Change-Id: If539ad9c1eb789f5b5f2b1454f7e8de6ee09b76c
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/30916
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: b7414a523b2dafa4de85d1b0cce0f1f58dd2a484