iwlwifi: nvm: fix VHT capability antenna-dependent fields [BUGFIX]
As the antenna dependent fields depend on the firmware file and not
the NVM, use those fields in the VHT capability creation.

Additionally, fix the STBC and antenna pattern consistency fields.

Change-Id: I4aa8ce8126c85eedc9bddc9a17e6f21ef489dff2
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
---
we merged the other patch too early, and now this code still uses
a field that's no longer read from the NVM and is zero ... as a
result, the VHT capabilities are wrong, particularly for WkP1.
Reviewed-on: https://gerrit.rds.intel.com/31015
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: 513952b5deeccb03d75b3bdac76783f3410283c3