mac80211: fix vif name tracing
If sdata doesn't have a valid dev (e.g. in case of monitor
vif), the vif_name field was initialized with (a length of)
some short string, but later was set to a different,
potentially larger one.

This resulted in out-of-bounds write, which usually
appeared as garbage in the trace log.

Simply trace sdata->name, as it should always have the
correct name for both cases.

Change-Id: I1ac36e3e5f7134601bdf24e6ba16d2469aaea452
Signed-off-by: Eliad Peller <eliadx.peller@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/31755
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: 3a9c0f64e8cdcd0d3ed3cd7e8b4f7dc706730033