cfg80211/mac80211: ignore signal if the frame was heard on wrong channel [BUGFIX]
On 2.4Ghz band, the channels overlap since the delta
between different channels is 5Mhz while the width of the
receiver is 20Mhz (at least).

This means that we can hear beacons or probe responses from
adjacent channels. These frames will have a significant
lower RSSI which will feed all kinds of logic with inaccurate
data. An obvious example is the roaming algorithm that will
think our AP is getting weak and will try to move to another
AP.

In order to avoid this, update the signal only if the frame
has been heard on the same channel as the one advertised by
the AP in its DS / HT IEs.
We refrain from updating the values only if the AP is
already in the BSS list so that we will still have a valid
(but inaccurate) value if the AP was heard on an adjacent
channel only.

To achieve this, stop taking the channel from DS / HT IEs
in mac80211. The DS / HT IEs is taken into account to
discard the frame if it was received on a disabled channel.
This can happen due to the same phenomenon: the frame is
sent on channel 12, but heard on channel 11 while channel
12 can be disabled on certain devices. Since this check
is done in cfg80211, stop even checking this in mac80211.

Change-Id: Idd71b1758e25a94d6b47391a5ab829a75314fc15
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
---
MWG100198324
Reviewed-on: https://gerrit.rds.intel.com/32047
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: b18395011a705a39fd68691af84ecedf1d33cf08