iwlwifi: mvm: BT Coex - enable per-AC BT priority
We can now define the priority against BT per AC. This is
possible with a newer firmware that allows to define the
priority with 2 bits.
Note that this change is compatible with older firmware
since older firmware will simply ignore the new bit (11),
and we still set the old bit (12) in the same cases as
before.

Change-Id: I4ae0ddccb31419fc7152235114fbbeeeb4cac144
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/32388
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: a84f1ec84882d5067971bf1cd68cc6208deab035