iwlwifi: testmode: fix dram hb end register value for dbgc
The register that holds the end address of the allocated
DRAM buffer for the DBGC in SfP is different than the one
for the DBGM on WkP in that it points to the last 256 byte
chunk that the DBGC is still *allowed* to write to.

Therefore, the address written to that register when the
DBGC is being configured needs to account for this allowed
writing or a buffer overrun by the data the DBGC is writing
will happen.

Change-Id: I1bbfd66b25caafba776c044c64ecd6d07e5472df
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/34721
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 72e211149d2c93caf7b823323642e6f2629979f3