[NOUPSTREAM] mac80211: don't try to copy the new chanctx to vlans for non-AP vifs
We call __ieee80211_vif_copy_chanctx_to_vlans() for all vif types when
using a new reserved chanctx.  This is harmless since the function
will just return if the vif type is not AP, but it causes a WARN to be
issued.

Fix this by only calling the function if the vif type is AP.

Change-Id: Ic421edd901d3eb9cb3aaaa059741790cc2a212c8
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/33280
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: 70a58c63291b8a74089fd2f022911a163aadafcb