iwlwifi: mvm: BT Coex - don't send updates for old API
A new API for BT Coex is coming. When the new API is
advertised by the firwmare, send the initialising command
but don't send updates based on the notifications.

This means that the BT Coex behavior will be affected in
case a firmware with the old API is detected, but it won't
blow up.

Change-Id: I51be2ae5e0ee14da6fbb627064cb7183c6c39a42
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/34607
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: 80ca5853918fd7af9da8a0cc70f7d96ac1dad347