iwlwifi: mvm: BT Coex - prepare towards new API
A new API is coming. This new API is not backward
compatible. So we need to keep the old commands to be able
to work with the former API.
Move all the current code into a new file: coex_legacy.
If a firmware with the new API is detected, we currently
just bail out since the implementation of the new API will
come in future patches.

Change-Id: I9a015a14dda1c0361ca5614fcd718c9ff481ebea
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/39772
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: baf7dedf62e0f0e7ebf793ce623e416ec57a0f67