iwlwifi: mvm: rs: fix and cleanup rs_get_rate_action
Change the down/upscale decision logic a bit to be based
on different success ratio thresholds. This fixes the implementation
compared to the rate scale algorithm which was planned to yield
optimal results. Also fix a case where a lower rate wasn't explored
despite being a potential for better throughput.
While at it rewrite rs_get_rate_action to be more clear and clean.

Change-Id: Ie1af6e2c4aebd88cc46fe5317904cc4f444cb0af
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
---
This was done to match the TLC SAS table in 4.2.7.3
Reviewed-on: https://gerrit.rds.intel.com/34785
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 0d44964bad7c47808b1b2113084272da98efe3b1