iwlwifi: mvm: rs: reinit rs if no tx for a long time
After being idle for a long time (>5sec) the rs statistics
will be stale so we prefer to reset rs and start from legacy
rates again. This gives better results when the attenuation
increased signficantly (e.g. we got further from the AP) and
after a while we start Tx
Note that the first Tx after the idle period will still go out
in the old modulation and rate but this seemed a simpler approach
compared to adding a timer or modifying mac80211 for this.
The negative impact is negligble as we'll recover quickly.

Change-Id: Icacee998889a3b62caae31388a45f480b8fac80e
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
---
This matches TLC SAS 4.2.8
It should also help with measurements done in TvA tests
Reviewed-on: https://gerrit.rds.intel.com/34835
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: 1310f250a6003066ff2f0f0a5c3fac56f4db3f5a