iwlwifi: mvm: rs: overhaul rs_get_best_rate [BUGFIX]
rs_get_best_rate determines the optimal rate to try in a
new Tx column. Currently we were sometimes trying a too high
rate which would lead us to fail and avoid switching to the
new column despite it having a potential to be better.
Change the logic to try and find the first rate which would
exceed the current actual throughput or be more aggressive
if the success ratio is good.

Change-Id: Ib51a54654a86930753d97d276e6f5c0658d732a2
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
---
This better matches the TLC SAS.
Helps with some of the degradation points in MWG100202310
Reviewed-on: https://gerrit.rds.intel.com/35243
Tested-by: IWL Jenkins
Reviewed-by: Assaf Azulay <assaf.azulay@intel.com>

x-iwlwifi-stack-dev: e5fdea0a26df572c515842e8f878bc9cb89cd352