iwlwifi: don't disable SCD chain extension on newer devices
7000 device series have a fix for this hardware feature.
Stop disabling it, and get an improvement in Tx throughput.
This feature allows the scheduler to fetch more frames on
the fly while an A-MPDU is being built - which means that
we can get larger A-MPDU. This, of course, give an
improvement in the Tx throughput.

Change-Id: Ib8be29e0cd85290458fb914072c944d7777ad6ff
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/35225
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: 89ae80a15fe862320b005d1fbba95d6a409b2660