iwlwifi: mvm: rs: avoid TPC while in search cycle
TPC is a power optimization which should only be attempted when
we're in the "stay in column" state after we've locked onto an optimal
rate.

Change-Id: I64b08a67518240db1c90b5c21b2feb643d71b84a
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
---
This is described in the TLC SAS.
I have also seen cases where TPC being activated messed up the search cycle.
Reviewed-on: https://gerrit.rds.intel.com/35537
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-by: EliadX Peller <eliad@wizery.com>

x-iwlwifi-stack-dev: 7d1194044c41969aeafddb4a89714629c6324d21