iwlwifi: mvm: Disable uAPSD for a DCM client
A binding object associates virtual MACs to a Phy.
Binding object can ask (and get) air time, thus
allowing several MACs to share the medium.
A binding object can be in either one of 3 states:
1) standalone - A single active MAC
2) Different Channel Mode (DCM) - There are 2 active
   MACs or more (not necessarily just clients), but each
   has its own channel.
3) Same Channel Mode (SCM) - There are 2 active
   MACs or more (not necessarily just clients), that
   share the same channel.
This patch introduces an additional condition for enabling uAPSD:
uAPSD should be disabled for a client in DCM.
The reason for that is a low latency issue in Miracast.

Change-Id: I1e758c777dd5e3ce804b99c46472d56b58731cb8
Signed-off-by: Avri Altman <avri.altman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/34496
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: af35578030ca57b572d413c7981e8e7efa2ecb10