iwlwifi: mvm: don't send zero quota to the firmware
There are some cases where we can currently send zero quota
for a valid binding, e.g. if we update while an interface is
bound to a channel context but not yet acting as an AP.

Avoid this by reordering the checks.

Change-Id: I232e97f4d630cd9aa447ababb4ad0fac590de80a
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/37796
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-by: Luciano Coelho <luciano.coelho@intel.com>

x-iwlwifi-stack-dev: 91d3363f54b42fb78d2b188fb0ba3d13233838c9